Optimize split position for append write in dleaf2_write()
Append write is common operation, and if append write, split at center
is not optimal for leaf split.

So, this splits leaf at sentinel, if leaf is most right and key is >=
sentinel.

FIXME1: we should apply this for bnode too.
FIXME2: this changes split position only if most right leaf. Should we
split at sentinel by filling hole too?

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>